Added map functionality
Includes react-native-maps package, breaks web testing of the application.

Requires API key to be entered in the app.json file.

Map screen is operational, with the ability to go to the specific test businesses displayed on the map.

Includes small changes to the BusinessDetailsScreen.js implemented to fix a bug. Doesn't change logic/architecture.

Changes made to the business data requires change of MapScreen.js as well. The logic isn't in its final form, will be updated.